requirements.txt
Flask>=2.2.0
Flask-Login>=0.6.0
Werkzeug>=2.2.0
python-dotenv>=0.20.0
openai>=0.28.0
whitenoise>=6.2.0
python-decouple>=3.6
gunicorn>=20.1.0
Jinja2>=3.1.0
itsdangerous>=2.1.0
click>=8.1.0
Flask-SQLAlchemy>=3.0.0
psycopg2-binary>=2.9.5
Flask-Migrate>=4.0.0
